requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
debtcollector==2.0.1
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-injector==0.1.1
django-model-utils==4.0.0
django-redis==4.11.0
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
injector==0.18.3
iso8601==0.1.12
Jinja2==2.11.2
jinja2schema==0.1.4
MarkupSafe==1.1.1
msgpack==1.0.0
mysqlclient==1.4.6
netaddr==0.7.19
netifaces==0.10.9
os-service-types==1.7.0
pbr==5.4.5
pyparsing==2.4.7
python-dotenv==0.13.0
python-http-client==3.2.7
python-magic==0.4.15
pytz==2020.1
PyYAML==5.3.1
redis==3.5.0
requests==2.23.0
rfc3986==1.4.0
sendgrid==6.3.0
six==1.14.0
sqlparse==0.3.1
stevedore==1.32.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
